requirements.txt
Django==2.0.2
django-bootstrap3==9.1.0
django-cleanup==1.0.1
psycopg2==2.7.3.2
pytz==2018.3
Pillow==5.2.0
xhtml2pdf==0.2.2
factory_boy== 2.11.1
Faker==1.0.1
